Разбиение диалога создания карточки на компоненты
