+++ src/app/(dashboard)/admin/schedule/_components/schedule-event-editor-dialog-steps/constants.ts
@@ -1,0 +1,9 @@
+export const DAY_OPTIONS = [
+  { value: "1", label: "Понедельник" },
+  { value: "2", label: "Вторник" },
+  { value: "3", label: "Среда" },
+  { value: "4", label: "Четверг" },
+  { value: "5", label: "Пятница" },
+];
+
+export const NONE_VALUE = "none";
